pyproject.toml
[project]
name = "sepsis-retrospective-email"
version = "0.1.0"
description = "Automated email system for sending personalized retrospective CDC Adult Sepsis Event (ASE) data to healthcare providers for quality assurance and improvement"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "polars>=0.19.0",
    "sqlalchemy>=2.0.0",
    "great-tables>=0.8.0",
    "python-dotenv>=1.0.0",
    "pyodbc>=5.0.0",
    "read_ad"
]

